+++ src/main/java/com/prography/lighton/artist/admin/application/ArtistApplicationQueryUseCase.java
@@ -1,7 +1,7 @@
 package com.prography.lighton.artist.admin.application;
 
-import com.prography.lighton.artist.admin.presentation.GetArtistApplicationDetailResponseDTO;
-import com.prography.lighton.artist.admin.presentation.GetArtistApplicationListResponseDTO;
+import com.prography.lighton.artist.admin.presentation.dto.response.GetArtistApplicationDetailResponseDTO;
+import com.prography.lighton.artist.admin.presentation.dto.response.GetArtistApplicationListResponseDTO;
 import com.prography.lighton.artist.common.domain.entity.enums.ApproveStatus;
 import java.util.List;
