+++ src/main/java/com/prography/lighton/artist/admin/application/ManageArtistApplicationService.java
@@ -1,6 +1,6 @@
 package com.prography.lighton.artist.admin.application;
 
-import com.prography.lighton.artist.admin.presentation.ManageArtistApplicationRequestDTO;
+import com.prography.lighton.artist.admin.presentation.dto.request.ManageArtistApplicationRequestDTO;
 import com.prography.lighton.artist.common.domain.entity.Artist;
 import com.prography.lighton.artist.common.domain.entity.enums.ApproveStatus;
 import com.prography.lighton.artist.users.infrastructure.repository.ArtistRepository;
