+++ src/main/java/com/prography/lighton/artist/admin/presentation/AdminController.java
@@ -2,6 +2,9 @@ package com.prography.lighton.artist.admin.presentation;
 
 import com.prography.lighton.artist.admin.application.ArtistApplicationQueryUseCase;
 import com.prography.lighton.artist.admin.application.ManageArtistApplicationUseCase;
+import com.prography.lighton.artist.admin.presentation.dto.request.ManageArtistApplicationRequestDTO;
+import com.prography.lighton.artist.admin.presentation.dto.response.GetArtistApplicationDetailResponseDTO;
+import com.prography.lighton.artist.admin.presentation.dto.response.GetArtistApplicationListResponseDTO;
 import com.prography.lighton.artist.common.domain.entity.enums.ApproveStatus;
 import com.prography.lighton.common.annotation.AdminOnly;
 import com.prography.lighton.common.utils.ApiUtils;
