--- src/main/java/com/prography/lighton/artist/admin/presentation/ManageArtistApplicationRequestDTO.java
@@ -1,4 +1,0 @@
-package com.prography.lighton.artist.admin.presentation;
-
-public record ManageArtistApplicationRequestDTO(String status) {
-}
