+++ src/main/java/com/prography/lighton/artist/admin/presentation/dto/request/ManageArtistApplicationRequestDTO.java
@@ -1,0 +1,4 @@
+package com.prography.lighton.artist.admin.presentation.dto.request;
+
+public record ManageArtistApplicationRequestDTO(String status) {
+}
